Implemented customer and single logged in customer

+++ src/main/java/com/cometbites/model/Card.java
@@ -10,8 +10,7 @@ public abstract class Card {
 
     private String cardName;
 
-    //TODO max int problem: 2147483647
-	public abstract int getNumber();
+	public abstract long getNumber();
 	
 	public String getId() {
 		return id;
